Extract MemberPage to its own file
This is in preparation of testing MemberPage with scraper_test
