+++ scraper.rb
@@ -7,6 +7,8 @@ require 'pry'
 require 'scraped'
 require 'scraperwiki'
 
+require_rel 'lib'
+
 # require 'open-uri/cached'
 # OpenURI::Cache.cache_path = '.cache'
 require 'scraped_page_archive/open-uri'
@@ -16,48 +18,6 @@ class MemberListPage < Scraped::HTML
     noko.css('div#content-core div.tileItem').drop(1).map do |p|
       p.at_css('h3 a/@href').text
     end
-  end
-end
-
-class MemberPage < Scraped::HTML
-  field :id do
-    url.split('/').last
-  end
-
-  field :name do
-    noko.at_css('h1#parent-fieldname-title').text.strip
-  end
-
-  field :area do
-    noko.at_css('div.electe span').text.strip
-  end
-
-  field :party do
-    noko.at_css('div.grup span').text.strip
-  end
-
-  field :email do
-    noko.at_css('div.email span').text.strip
-  end
-
-  field :birth_date do
-    Date.parse(noko.at_css('div.datanaixement span').text.strip).to_s
-  end
-
-  field :image do
-    noko.at_css('div.foto img/@src').text
-  end
-
-  field :election_date do
-    Date.parse(noko.at_css('div.dataeleccio span').text.strip).to_s
-  end
-
-  field :term do
-    2015
-  end
-
-  field :source do
-    url
   end
 end
